requirements/base.txt
numpy==1.24.3
torch==2.1.0
gymnasium==0.29.1
pybullet==3.2.5
stable-baselines3==2.1.0
tensorboard==2.14.0
PyYAML>=6.0.1